final version of irishspell
